ApplyMigrations and AddTenantAndApplyMigrations return types.MigrationResult which contains statistics about newly applied DB migrations
