+++ db/db_mssql_test.go
@@ -1,0 +1,122 @@
+package db
+
+import (
+	"testing"
+
+	"github.com/lukaszbudnik/migrator/config"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestDBCreateDialectMSSQLDriver(t *testing.T) {
+	config := &config.Config{}
+	config.Driver = "sqlserver"
+	dialect, err := newDialect(config)
+	assert.Nil(t, err)
+	assert.IsType(t, &msSQLDialect{}, dialect)
+}
+
+func TestMSSQLGetMigrationInsertSQL(t *testing.T) {
+	config, err := config.FromFile("../test/migrator.yaml")
+	assert.Nil(t, err)
+
+	config.Driver = "sqlserver"
+
+	dialect, err := newDialect(config)
+	assert.Nil(t, err)
+
+	insertMigrationSQL := dialect.GetMigrationInsertSQL()
+
+	assert.Equal(t, "insert into migrator.migrator_migrations (name, source_dir, filename, type, db_schema, contents, checksum) values (@p1, @p2, @p3, @p4, @p5, @p6, @p7)", insertMigrationSQL)
+}
+
+func TestMSSQLGetTenantInsertSQLDefault(t *testing.T) {
+	config, err := config.FromFile("../test/migrator.yaml")
+	assert.Nil(t, err)
+
+	config.Driver = "sqlserver"
+	dialect, err := newDialect(config)
+	assert.Nil(t, err)
+	connector := baseConnector{config, dialect, nil}
+	defer connector.Dispose()
+
+	tenantInsertSQL := connector.getTenantInsertSQL()
+
+	assert.Equal(t, "insert into migrator.migrator_tenants (name) values (@p1)", tenantInsertSQL)
+}
+
+func TestMSSQLDialectGetCreateTenantsTableSQL(t *testing.T) {
+	config, err := config.FromFile("../test/migrator.yaml")
+	assert.Nil(t, err)
+
+	config.Driver = "sqlserver"
+
+	dialect, err := newDialect(config)
+	assert.Nil(t, err)
+
+	createTenantsTableSQL := dialect.GetCreateTenantsTableSQL()
+
+	expected := `
+IF NOT EXISTS (select * from information_schema.tables where table_schema = 'migrator' and table_name = 'migrator_tenants')
+BEGIN
+  create table [migrator].migrator_tenants (
+    id int identity (1,1) primary key,
+    name varchar(200) not null,
+    created datetime default CURRENT_TIMESTAMP
+  );
+END
+`
+
+	assert.Equal(t, expected, createTenantsTableSQL)
+}
+
+func TestMSSQLDialectGetCreateMigrationsTableSQL(t *testing.T) {
+	config, err := config.FromFile("../test/migrator.yaml")
+	assert.Nil(t, err)
+
+	config.Driver = "sqlserver"
+
+	dialect, err := newDialect(config)
+	assert.Nil(t, err)
+
+	createMigrationsTableSQL := dialect.GetCreateMigrationsTableSQL()
+
+	expected := `
+IF NOT EXISTS (select * from information_schema.tables where table_schema = 'migrator' and table_name = 'migrator_migrations')
+BEGIN
+  create table [migrator].migrator_migrations (
+    id int identity (1,1) primary key,
+    name varchar(200) not null,
+    source_dir varchar(200) not null,
+    filename varchar(200) not null,
+    type int not null,
+    db_schema varchar(200) not null,
+    created datetime default CURRENT_TIMESTAMP,
+		contents text,
+		checksum varchar(64)
+  );
+END
+`
+
+	assert.Equal(t, expected, createMigrationsTableSQL)
+}
+
+func TestMSSQLDialectGetCreateSchemaSQL(t *testing.T) {
+	config, err := config.FromFile("../test/migrator.yaml")
+	assert.Nil(t, err)
+
+	config.Driver = "sqlserver"
+
+	dialect, err := newDialect(config)
+	assert.Nil(t, err)
+
+	createSchemaSQL := dialect.GetCreateSchemaSQL("def")
+
+	expected := `
+IF NOT EXISTS (select * from information_schema.schemata where schema_name = 'def')
+BEGIN
+  EXEC sp_executesql N'create schema def';
+END
+`
+
+	assert.Equal(t, expected, createSchemaSQL)
+}
