+++ db/db_postgresql_test.go
@@ -1,0 +1,45 @@
+package db
+
+import (
+	"testing"
+
+	"github.com/lukaszbudnik/migrator/config"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestDBCreateDialectPostgreSQLDriver(t *testing.T) {
+	config := &config.Config{}
+	config.Driver = "postgres"
+	dialect, err := newDialect(config)
+	assert.Nil(t, err)
+	assert.IsType(t, &postgreSQLDialect{}, dialect)
+}
+
+func TestPostgreSQLGetMigrationInsertSQL(t *testing.T) {
+	config, err := config.FromFile("../test/migrator.yaml")
+	assert.Nil(t, err)
+
+	config.Driver = "postgres"
+
+	dialect, err := newDialect(config)
+	assert.Nil(t, err)
+
+	insertMigrationSQL := dialect.GetMigrationInsertSQL()
+
+	assert.Equal(t, "insert into migrator.migrator_migrations (name, source_dir, filename, type, db_schema, contents, checksum) values ($1, $2, $3, $4, $5, $6, $7)", insertMigrationSQL)
+}
+
+func TestPostgreSQLGetTenantInsertSQLDefault(t *testing.T) {
+	config, err := config.FromFile("../test/migrator.yaml")
+	assert.Nil(t, err)
+
+	config.Driver = "postgres"
+	dialect, err := newDialect(config)
+	assert.Nil(t, err)
+	connector := baseConnector{config, dialect, nil}
+	defer connector.Dispose()
+
+	tenantInsertSQL := connector.getTenantInsertSQL()
+
+	assert.Equal(t, "insert into migrator.migrator_tenants (name) values ($1)", tenantInsertSQL)
+}
